Fix a race condition when adding tap interfaces to bridge
We use ensure_tap_in_bridge to make sure a tap interface is added to a
Linux kernel bridge.

However, this hangs on occasion when called from the device monitor
callback.

Working on the assumption that this is a deadlock (device monitor
listens for network device events, and on receiving one does work that
causes a network device event it's currently not listening for), split
the device monitor into front/back tasks so that the device monitor may
continue to listen while the callback is done.

Closes-Bug: #1825404
Change-Id: I14114e89103ceb990c3e291fd70ca75067fbb4ff