Fix eventlet thread for bringing tap into Linux Bridge
Today, the only function which is draining the self.external_devices
queue is maybe_ensure_external_devices(), and
maybe_ensure_external_devices() is only called when there is new port
being added to etcd.

This means that, when eventlet detected a new tap interface in Linux,
the interface will be added to the queue via
_consider_external_device(), but nothing will happen after that for
an indefinite period of time.

This adds an independent eventlet to do the processing on demand.

Change-Id: I129e211e199434a2ab104ec42bace8a408e3a0ea