Prevent a race condition when loading the trunk plugin
When the trunk plugin loads the etcdAgentCommunicator,
it sets the bind notifier to an empty lambda function. This
causes a race condition in the etcdAgentCommunicator object
used by mech_vpp to send port bind complete requests.
This patch resolves the issue by using a separate
JournalManager class that does not start any watcher
threads.

Closes-Bug: #1845537
Change-Id: I209b31b3e24dbb9bd6afeb327f15fe9883a57b71