Fix for incorrect DHCP tap processing
When the neutron DHCP agent lags behind the ML2 agent slightly in the
DHCP tap creation process, it brings into play the device monitor
whose job is to monitor for new tap interfaces and pass along the
information to another thread which plumbs the tap interface under
the right kernel bridge.

A bug in the way the list of tap interfaces is processed causes
failure of addition of the tap under the kernel bridge which
ultimately results in the instances not being able to obtain their
DHCP IP addrs.

Change-Id: Iaa238a497bc06196b04d05edce46547bb8785bf3
Closes-Bug: #1854165