Fix for sw_interface_details change in 19.08.1
In VPP 19.08.1, the interface_name and tag fields in sw_interface_details
have changed from the earlier u8/bytes type to the new string type. This
causes the ML2 agent to die with the following error messagae with
python3:

'str' object has no attribute 'decode'

This is because on python3, PAPI converts string type to the python
str type which is unicode by default and hence there's no decode
attribute.

Change-Id: Ic595a058cf20b2a7ea661b393ed16dd0a7e562a9
Closes-Bug: #1855302