Rotate bind address of AF_NETLINK socket
Upstream py3x tests seem to have difficulties with an address clash
on the device listener.  While this doesn't identify the root cause,
it's good practice not to try repeating an operation that will never
work and this ensures we try more than one address if we have a problem.

Also: enable the py35 test as a voting job and fix a small requirements
issue.  (This demonstrates this helps.)  And add openstack-python-jobs
to .zuul.conf because we're the first merge after removing it from
project-config.

Change-Id: I474a5d57eac9def5adb43120f1d70907460c1568