Change monkey patch behaviour
We have found that the rootwrap code now uses threading.Lock for
locking.  We were previously not patching threading.Lock because we had
need of real thread locks (for interacting with VPP callbacks) and
didn't use eventlet locks, but in this instance we must monkey patch
threading using eventlet for a happy life.

Preserve the reference to the original threading.Lock type before monkey
patching; pass this explicitly to VPP as an optional argument so that it
can avoid the problems eventlet causes.  Neutron libraries can now use
Locks that work between eventlets.

Change-Id: Ib728114e8dca637bdf5ba6b782fdabd6531e6632