Deal with poorly formed secrity group subnets
1.2.3.4/0 is completely acceptable in a security group rule.
Previously, VPP accepted this in its API calls but this is not
currently the case.

Fix up at the high level so that e.g. 1.2.3.4/8 == 1.0.0.0/8 (which is
acceptable, since 1.0.0.0/8 contains the network address).  So
effectively we treat etcd as a loose definition and we fix it
immediately we consume it so that it is a network mask throughout
networking-vpp from then on.

Change-Id: I21430591bb0a1ce2b86e93bb405015001557953f