Update SG mechdriver code to add rules properly
Adding a rule via the API adds the rule on the next SG update,
due to the weird way in which Neutron updates its database.  This
results in intermittent cases where a rule is added but not pushed out
to VPP until the next rule is added or the SG is changed (e.g. a name
change).  It's not true consistently, but it has been seen in test.

Use the opportunity to tidy the code, which is overengineered to
assume multiple-object updates when that never in fact happens.

Change-Id: I89b7ece88edf4613ccdbb5c14cae4b81608b483c