Fix for RuntimeError: dictionary changed size during iteration
In py3 trying to modify a mutable object while iterating over it produces
the following error:

    RuntimeError: dictionary changed size during iteration

but we do this as we process the subport/trunk port relationship.
Get a static copy of our work items out of the dict beforehand so that
we can update the dict as we go.

Change-Id: Iaa1937d5ef3ac683b5f0085f5275e47418ed7eb8
Closes-Bug: #1882672