Fix use of mock
For ourselves, we know that 'mock' is now available in py33 and better
as unittest.mock, so we have converted our inputs.

For Neutron, we use neutron.tests but don't have it as a dependency,
and it wants to use the 'mock' package as-is, so we include 'mock'
as a test requirement.  This should probably be checked in other
Neutron driver repos to work out why we're missing the requirement;
there's probably something in our CI files that causes the problem.
In the meantime this is safe, if slightly untidy.

Change-Id: I9f13d6f3b2d46fae83483148948cf02b430c9b15