mech_vpp validates by the list of configured network type drivers
This will, on an all-in-one node, be whatever the agent is loading;
in multi-node configuations, network_drivers should be consistent
between control and compute nodes.

Since the valid network types that a mechanism driver accepts are
precisely dependent on the ones the agent supports, this allows
them to be configured to work in tandem.

Change-Id: I0f698b64fbb4a799dcf134d2b5bf3bf516d0dade