COMPAT CLEANUP: remove conditional imports
Remove all the conditionals to create the right imports in compat.py
where they relate to unsupported versions.

This is, in fact, most of them; the actual import lines will be
fixed in a subsequent patch.

Changed use of n_const and plugin_constants in the main body of the
code because they were being mixed up, making for harder work.  The
reason was that at some point the plugin types were moved to a
different file, confusing matters; L3 is a plugin constant, Neutron-
inbuilt network types are not.

We should also look for use of TYPE_VXLAN as it's quite likely that
we shouldn't be using it to refer to GPE networks at all.

Change-Id: I96356a3500b3bbc90f6acb2673fe75df5d027468