COMPAT CLEANUP: stop importing things from compat
Compat is for compatibility; there are certain things that no longer
need help because they're consistent across all OpenStack versions we
support (really Queens+ at this point).  Remove the compat.py code
that supports this and stop using compat in import lines.

Also removed some conditional importing that was happening outside of
the compatibility layer - which shouldn't happen, but did in odd
places.

Change-Id: I913b184713a85b05ff97dcc14c9459025cb0c0e5