vxlan-gpe: fix agent crash due to lisp_add_del_local_eid API call failure
Before trying to add a new lisp local eid in VPP, ML2 agent
checks against the in-memory map of local eids. But when the agent
is restarted, it builds the map of local eids by querying VPP. This
results in the in-memory map to contain MACAddress object from PAPI
instead of a string. This causes the check to fail and the attempt
to create the local eid fails causing the agent to crash.

Change-Id: Ib342260407100f6aa3e53a531f2130be9d4a9d83
Closes-Bug: #1892749