vxlan-gpe: lisp l2 arp entry fails to delete
When the gpe listener receives a gpe port delete event from etcd, it
deletes the remote mapping entry for the port as well the lisp l2
arp entry. Before deleting the l2 arp entry, it checks for the
existence of the entry in VPP. This check fails because the IP address
returned by PAPI is the Python IPv4Address object whereas the gpe
listener expects the binary address.

Change-Id: I3935daf12c5b96fad56c46c970580582ad47f8a4
Closes-Bug: #1893766