Fix an ifidx!=0 test
loopback_idx == 0 is a valid possiblility, so a test should have read
loopback_idx is not None (i.e. there is no loopback interface).
However, the mocking in the test repo thinks that that 'None' return
value should actually be 'False', which means that correcting the check
in the code breaks the test.

Change-Id: I05310b0547248b364778409e518fc25db792ccf7