Increase type checking of server.py
Server.py has a lot of functions in, not all of which have signatures,
so we can't go full-on with type checking, but we can at least make it
check the bodies of all its functions.

Note the spot where bvi_if_idx can be None ('not found') - the checks
would rule out 0 (valid ifidx) and in some cases aren't made (go do
things with None that are unacceptable).  All patched up.

Change-Id: Ib90d3e3e141379ec1729dc8218e0be687b95069e