Add typing to macip ACL fetching
We're still using PAPI's return, but we assign a type to it which
makes it explicit what we're getting and what callers can do with it.

'get_macip_acl_dumo' is a redundant and cumbersome name; _dump is
more PAPI parlance, so I went with get_macip_acls.

Removed the proxy function, it only makes things less clear.

Change-Id: Ib9d9a888d734ef7f8fbe23475c04ac75029014cd